Add files via upload
The theme of this project is intelligent visual surveillance systems. In recent times, we used surveillance cameras for monitoring and recording moments, but manual surveillance and real-time monitoring is one of the most important and challenging branches of computer vision, which has been widely applied in people’s  life, such as monitoring security. The presence of surveillance cameras and a warning sign indicating that the area is under monitoring can serve as a significant deterrent to criminals and thieves, as the recorded footage can be used to identify people and trace their activities. It can be more advanced with Wi-Fi, which is a local area network running in a local environment or in a distributed setting. Wi-Fi network protocol is one of the leading communication technologies used in the IoT world which supports low transmit power along with low cost. ESP32 is the second generation of Express if corporation IoT solution and it includes Wi Fi. ESP32 reduces high network traffic and computing load. This system facilitates the user to receive notifications whenever the intrusion is detected with the help of sensors connected with the surveillance cameras.
By this project, we are implementing a smart surveillance system using ESP-32 CAM module. So, by using this system, if an intruder is entered into the home or any suspects were walking around your home one can get an immediate alert to their mobile through SMS along with-it buzzer will generate an alarm. This system is implemented based on Arduino UNO that processes and detects the presence of an intrusion Captured pictures are sent to the registered Email account. As a result, we may infer that our system is capable of achieving all of the aforementioned goals and of overcoming the existing system's challenges. With our proposed system, surveillance is vastly improved. The designed system enabled us to achieve the following goals: real-time monitoring, reduced human intervention, and use of active sensors in the field.